+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add external trigger to `Modal` component.
 - Feature: Add responsive mixin `k-u-margin(…)`.
 - Feature: Add `Loader` on `ProjectSimilarCard` component.
+- Feature: Add `leftArrowDisabled`, `rightArrowDisabled`, `onLeftArrowClick` and
+  `onRightArrowClick` props to `ProjectSimilarCard` component.
+- Feature: Add `SimilarProjectsCard` component.
 
 ## [13.5.0] - 2017-07-28